Adding four new Ansible Modules
The first module gets facts about all hosts in Rubrik. Currently we're using this internally to find any machines that don't have a vm registered.

The second module refreshes the specified vcenter environment on request. We use this during VM provisioning or decom processes.

The third module registers a VM's agent in Rubrik so the agent can run backups.

The fourth module gets facts for all of the vcenter environments.